Adding -monsterevade and updating relevant flags
This update adds in -monsterevade, a flag that "fixes" the vanilla
"bug" where monsters do not load evade (physical or magical) at the
beginning of battle; it is strictly a difficulty modifier.
(The solution also tweaks magic behaviour; now monsters ignore
magic defense and evade when casting spells on themselves, unless the
magic defense is 255.)

This commit also adds in uispec for the ATB kit, since I forgot to
do that earlier.