Changing treasure weighting under vanilla:miabs
Locations with miabs are, in this version of FE, given better
treasure distributions than locations without miabs. However, on
vanilla:miabs these locations are much easier to loot, so we can
adjust the weights with a matrix similarly to how Twildish boosts
Tpro weights. The new weights are not quite "back to Tpro" but
they result in similar curves.