Adding Tplayable and -smith:playable (and flagspec)
Tplayable is essentially the "Falcon Dive" flag; all chests will
contain items that at least one character available in the seed
will be able to use. Under Omnidextrous, this flag does nothing.
Under Fist Fight, the only weapons available are claws. This flag
does not apply to KI check zonk rewards or MIABs. -smith:playable
restricts the smith item, if it is not the vanilla Excalbur, to be
usable by at least one character available in the seed (including
if it is not a supersmith weapon).

This commit also adds in some uispec/uispec_flagdescriptions for
Bwoahdin and Kunweighted which did not make it into a previous
commit.