Adding -monsterflee
At one point, the original devs had intended for monsters to be able
to run away from you in battle. All of the code is still there, but
they changed three things so that the monster Flee command would
never be used. This flag undoes those changes and lets monsters flee
again.

In particular, the check for commands outside a certain range is
rewritten to allow for Flee, the process that sets Flee also now sets
self-targetting, and monsters have evade stats to compare against.
(This flag requires that -monsterevade be turned on.)

Finally, for fun monsters are allowed to run from battle even if your
party can't, as a treat for using Porom's Cry mechanic.