Adding -exp:geometric
Normally each monster of the same type killed in battle yields the
same experience; this flag makes it so each subsequent monster of the
same type gives 9/10 as much of the experience as the previous one,
meaning repeated grinds (D$, MacGs, etc.) have diminishing returns.
Total experience is capped at ten times the original monster type exp,
thanks to the formula for geometric sums.

The changes involve a new replica 16-bit division function in util.f4c
and another util.f4c function that does the geometric scaling. If
desired, the scale factor could be changed to be something else, but
9/10 seemed reasonable.