Adding -exp:objectivebonus
This flag provides a bonus 12.5% experience (rounded down) in battle
for each objective completed when objectives are enabled (and is
turned off under Onone). There is no cap, so if you do e.g. 24
objectives, you'll get roughly a 300% boost to your experience
earnings. The goal of this flag was to provide a reason to do
objectives first instead of grinding as soon as possible.

The specific choice of 12.5% (1/8) was chosen as the closest power of
1/2 that was near 10%, to allow bit-shifting. Picking different
fractions, or trying to recover the missing experience, will be a more
challenging implementation. Technically, this boost happens after the
key item bonus, so potentially less experience is missed due to
rounding in that case.

In the future, it might be worthwhile to make a Util.f4c function
that counts objectives; currently the objective counting script in
tracker.f4c is duplicated in experience_acceleration.f4c.

A typo in the flag description for -exp:geometric was also corrected.