add ability to mark a flag as coming from a fork
to mark a flag as a fork, add a ^ to the flag in the uispec.txt file and run the appropriate compile_all_scripts file to regenerate the generated files.

also updates fork icon to git icon, uses fork icon for marking a specific flag as a fork.
add css for h4 and h5 in fork_info.css