Small code change for Zeromus rando
I'd written the conditions for the Dart and Aim reactions with opcode
$08, which (I think) was intended for checking if a monster had been
hit with a type of attack in the past, and then changing their
attacking script going forward. That's not really intended for
reactions. I've changed the two conditions to use opcode $07 instead,
which works for all characters using parameter $19 (like all of the
other "react to ___" conditions in the vanilla game do). This should
prevent anything funky from happening (which I hadn't previously seen,
but better safe than sorry).